third_party/blink/web_tests/svg/dynamic-updates/SVGFETurbulenceElement-svgdom-stitchTiles-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stitchTiles' property of the SVGFETurbulenceElment object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS turbulence.stitchTiles.baseVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS turbulence.stitchTiles.baseVal is SVGFETurbulenceElement.SVG_STITCHTYPE_NOSTITCH
PASS successfullyParsed is true

TEST COMPLETE
